pyproject.toml
[project]
name = "palia-checklist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["mwclient>=0.11.0", "python-dotenv>=1.1.1", "rich>=13.7.1"]
